+++ crypto_kem/mceliece6960119f/clean/decrypt.c
@@ -77,7 +77,7 @@ int PQCLEAN_MCELIECE6960119F_CLEAN_decrypt(unsigned char *e, const unsigned char
 
     //
 
-    check = w;
+    check = (uint16_t)w;
     check ^= SYS_T;
 
     for (i = 0; i < SYS_T * 2; i++) {
