+++ crypto_kem/mceliece8192128/clean/decrypt.c
@@ -75,7 +75,7 @@ int PQCLEAN_MCELIECE8192128_CLEAN_decrypt(unsigned char *e, const unsigned char 
 
     //
 
-    check = w;
+    check = (uint16_t)w;
     check ^= SYS_T;
 
     for (i = 0; i < SYS_T * 2; i++) {
